requirements.txt
fastapi
uvicorn
pymupdf
gradio
requests
huggingface_hub[hf_xet]
transformers
sentence-transformers
faiss-cpu
torch